fix: refactor tests, cleanup and move to directories and individual files
